Feature/custom server validation

+++ DbNetSuiteCore/Services/FormService.cs
@@ -10,8 +10,9 @@ using DbNetSuiteCore.Constants;
 using DbNetSuiteCore.Enums;
 using DbNetSuiteCore.Extensions;
 using MongoDB.Bson;
-using System.Text.RegularExpressions;
-using NUglify.Helpers;
+using Microsoft.Extensions.Options;
+using DbNetSuiteCore.Middleware;
+using DocumentFormat.OpenXml.Spreadsheet;
 
 namespace DbNetSuiteCore.Services
 {
@@ -21,7 +22,7 @@ namespace DbNetSuiteCore.Services
         {
         }
 
-        public async Task<Byte[]> Process(HttpContext context, string page)
+        public async Task<Byte[]> Process(HttpContext context, string page, IOptions<DbNetSuiteCoreOptions> options)
         {
             try
             {
@@ -29,7 +30,7 @@ namespace DbNetSuiteCore.Services
                 switch (page.ToLower())
                 {
                     case "formcontrol":
-                        return await FormView();
+                        return await FormView(options);
                     default:
                         return new byte[0];
                 }
@@ -41,7 +42,7 @@ namespace DbNetSuiteCore.Services
             }
         }
 
-        private async Task<Byte[]> FormView()
+        private async Task<Byte[]> FormView(IOptions<DbNetSuiteCoreOptions> options)
         {
             FormModel formModel = GetFormModel();
             formModel.TriggerName = RequestHelper.TriggerName(_context);
@@ -51,11 +52,11 @@ namespace DbNetSuiteCore.Services
             switch (formModel.TriggerName)
             {
                 case TriggerNames.Apply:
-                    return await ApplyUpdate(formModel);
+                    return await ApplyUpdate(formModel, options);
                 case TriggerNames.Toolbar:
                     return await Toolbar(formModel);
                 case TriggerNames.Delete:
-                    return await ApplyDelete(formModel);
+                    return await ApplyDelete(formModel, options);
                 case TriggerNames.Insert:
                     return await InitialiseInsert(formModel);
                 default:
@@ -113,7 +114,7 @@ namespace DbNetSuiteCore.Services
         private List<object> PrimaryKeyValue(DataRow dataRow)
         {
             List<object> primaryKeyValues = new List<object>();
-            foreach (object? value in (dataRow.ItemArray ?? new object[] { })) 
+            foreach (object? value in (dataRow.ItemArray ?? new object[] { }))
             {
                 if (value == null)
                 {
@@ -128,17 +129,17 @@ namespace DbNetSuiteCore.Services
                     primaryKeyValues.Add(value);
                 }
             }
-           
+
             return primaryKeyValues;
         }
-        private async Task<Byte[]> ApplyUpdate(FormModel formModel)
+        private async Task<Byte[]> ApplyUpdate(FormModel formModel, IOptions<DbNetSuiteCoreOptions> options)
         {
             FormViewModel formViewModel = new FormViewModel(formModel);
             var committed = false;
 
             if (formModel.ClientEvents.Keys.Contains(FormClientEvent.ValidateUpdate) == false)
             {
-                if (await ValidateRecord(formModel))
+                if (await ValidateRecord(formModel, options))
                 {
                     await Commit();
                 }
@@ -146,7 +147,7 @@ namespace DbNetSuiteCore.Services
             }
             else if (formModel.ValidationPassed == false)
             {
-                formModel.ValidationPassed = await ValidateRecord(formModel);
+                formModel.ValidationPassed = await ValidateRecord(formModel, options);
             }
             else
             {
@@ -196,14 +197,21 @@ namespace DbNetSuiteCore.Services
             }
         }
 
-        private async Task<Byte[]> ApplyDelete(FormModel formModel)
+        private async Task<Byte[]> ApplyDelete(FormModel formModel, IOptions<DbNetSuiteCoreOptions> options)
         {
             try
             {
-                await DeleteRecord(formModel);
-                formModel.PrimaryKeyValues = new List<List<object>>();
-                formModel.Message = ResourceHelper.GetResourceString(ResourceNames.Deleted);
-                formModel.MessageType = MessageType.Success;
+                if (await CustomDelegateValidation(options.Value.FormDeleteValidationDelegate, formModel))
+                {
+                    await DeleteRecord(formModel);
+                    formModel.PrimaryKeyValues = new List<List<object>>();
+                    formModel.Message = ResourceHelper.GetResourceString(ResourceNames.Deleted);
+                    formModel.MessageType = MessageType.Success;
+                }
+                else
+                {
+                    throw new Exception(formModel.Message);
+                }
             }
             catch (Exception ex)
             {
@@ -221,36 +229,73 @@ namespace DbNetSuiteCore.Services
         private async Task<Byte[]> InitialiseInsert(FormModel formModel)
         {
             formModel.Mode = FormMode.Insert;
+            formModel.FormValues = new Dictionary<string, string>();
             await GetLookupOptions(formModel);
             return await View("Form/__Form", new FormViewModel(formModel));
         }
 
-        private async Task<bool> ValidateRecord(FormModel formModel)
+        private async Task<bool> ValidateRecord(FormModel formModel, IOptions<DbNetSuiteCoreOptions> options)
         {
-            if (ValidateErrorType(formModel, ResourceNames.Required))
+            var validationTypes = new List<ResourceNames>() { ResourceNames.Required, ResourceNames.DataFormatError, ResourceNames.MinCharsError, ResourceNames.MaxCharsError, ResourceNames.MinValueError, ResourceNames.PatternError };
+
+            PopulateGuidPrimaryKey(formModel);
+
+            foreach (var validationType in validationTypes)
             {
-                if (ValidateErrorType(formModel, ResourceNames.DataFormatError))
+                if (ValidateErrorType(formModel, validationType) == false)
                 {
-                    if (ValidateErrorType(formModel, ResourceNames.MinCharsError))
-                    {
-                        if (ValidateErrorType(formModel, ResourceNames.MaxCharsError))
-                        {
-                            if (ValidateErrorType(formModel, ResourceNames.MinValueError))
-                            {
-                                if (ValidateErrorType(formModel, ResourceNames.PatternError))
-                                {
-                                    if (await ValidatePrimaryKey(formModel))
-                                    {
-                                        return true;
-                                    }
-                                }
-                            }
-                        }
-                    }
+                    return false;
                 }
             }
 
-            return false;
+            if (await ValidatePrimaryKey(formModel) == false)
+            {
+                return false;
+            }
+
+            FormValidationDelegate? validationDelegate = formModel.Mode == FormMode.Update ? options?.Value.FormUpdateValidationDelegate : options?.Value.FormInsertValidationDelegate;
+            return await CustomDelegateValidation(validationDelegate, formModel);
+        }
+
+        private void PopulateGuidPrimaryKey(FormModel formModel)
+        {
+            if (formModel.Mode == FormMode.Update)
+            {
+                return;
+            }
+
+            foreach (FormColumn? formColumn in formModel.Columns.Where(c => c.PrimaryKeyRequired && c.DataType == typeof(Guid)))
+            {
+                var columnName = formColumn.ColumnName;
+
+                var value = string.Empty;
+
+                if (formModel.FormValues.ContainsKey(columnName) == false || string.IsNullOrEmpty(formModel.FormValues[columnName]))
+                {
+                    formModel.FormValues[columnName] = Guid.NewGuid().ToString();
+                }
+            }
+        }
+
+        private async Task<bool> CustomDelegateValidation(FormValidationDelegate? validationDelegate, FormModel formModel)
+        {
+            if (validationDelegate == null)
+            {
+                return true;
+            }
+
+            bool result = await validationDelegate(formModel, _context!, _configuration);
+
+            if (result == false)
+            {
+                if (string.IsNullOrEmpty(formModel.Message))
+                {
+                    formModel.Message = "Custom validation failed";
+                }
+                formModel.MessageType = MessageType.Error;
+            }
+
+            return result;
         }
 
         private async Task<bool> ValidatePrimaryKey(FormModel formModel)
@@ -262,7 +307,7 @@ namespace DbNetSuiteCore.Services
 
             if (await RecordExists(formModel))
             {
-                formModel.Columns.Where(c => c.PrimaryKeyRequired).ForEach(c => c.InError = true);
+                formModel.Columns.Where(c => c.PrimaryKeyRequired).ToList().ForEach(c => c.InError = true);
                 formModel.Message = ResourceHelper.GetResourceString(ResourceNames.PrimaryKeyExists);
                 formModel.MessageType = MessageType.Error;
                 return false;
@@ -275,6 +320,11 @@ namespace DbNetSuiteCore.Services
         {
             foreach (FormColumn? formColumn in formModel.Columns)
             {
+                if (formModel.Mode == FormMode.Update || formColumn.Required == false || formColumn.PrimaryKey)
+                {
+                    continue;
+                }
+
                 var columnName = formColumn.ColumnName;
 
                 var value = string.Empty;
@@ -286,10 +336,6 @@ namespace DbNetSuiteCore.Services
                 else if (formModel.Mode == FormMode.Insert && formColumn.PrimaryKeyRequired)
                 {
                     value = string.Empty;
-                }
-                else if (formModel.Mode == FormMode.Update || formColumn.Required == false || formColumn.PrimaryKey)
-                {
-                    continue;
                 }
 
                 ValidateFormValue(formColumn, value, resourceName, formModel);
@@ -322,9 +368,10 @@ namespace DbNetSuiteCore.Services
                 formModel.JSON = TextHelper.Decompress(RequestHelper.FormValue("json", string.Empty, _context));
                 formModel.CurrentRecord = formModel.ToolbarPosition == ToolbarPosition.Hidden ? 1 : GetRecordNumber(formModel);
                 formModel.SearchInput = RequestHelper.FormValue("searchInput", string.Empty, _context).Trim();
-                formModel.FormValues = RequestHelper.FormColumnValues(_context);
-                formModel.Columns.ToList().ForEach(column => column.InError = false);
+                formModel.FormValues = RequestHelper.FormColumnValues(_context, formModel);
+                formModel.Columns.ToList().ForEach(column => column.InError = false);   
                 formModel.Message = string.Empty;
+                formModel.Modified = RequestHelper.GetModified(_context, formModel); ;
                 formModel.ValidationPassed = ComponentModelExtensions.ParseBoolean(RequestHelper.FormValue("validationPassed", formModel.ValidationPassed.ToString(), _context));
                 formModel.CommitType = null;
                 formModel.SearchDialogConjunction = RequestHelper.FormValue("searchDialogConjunction", "and", _context).Trim();
